requirement.txt
#pure-python-adb
adb_shell
pyyaml
